Jwkoelewijn zookeeper namespace support (#89)
* Added a zookeeper 'namespace' option.

Zookeeper instances can be namespaces, meaning that all znodes are
prefixed with the same name. This means that broker information would be
found at /namespace/brokers/ids instead of /brokers/ids.

This feature branch adds support for this, defaulting to an empty
namespace. Specs have been added to test the functionality.

* chroot

* changelog, version dump and readme update